[refactor]#27 API와 DTO 패키지 구분
